Refactor: 그룹 멤버 관리 테이블 리팩토링
