+++ src/pages/group-management/member-management/components/index.ts
@@ -1,1 +1,1 @@
-export * from './member-table'
+export * from './member-table-section'
