+++ src/pages/group-management/member-management/components/member-table-section/index.ts
@@ -1,0 +1,1 @@
+export { GroupMemberTableSection } from './MemberTable'
