--- src/pages/group-management/member-management/components/member-table/group-table/index.ts
@@ -1,1 +1,0 @@
-export { GroupTable } from './GroupTable'
